+++ src/features/home/Home.tsx
@@ -1,9 +1,9 @@
 import { Link } from "react-router-dom";
-import { useWorkspaceStore } from "../../state/stores/workspace-store";
+import { useWorkspaceStore } from "@/state/stores/workspace-store";
 import { useWorkspaces } from "./workspace-query";
-import { routePaths } from "../../routes/route-paths";
-import { useCurrentUser } from "../../data/user/active-user-query";
-import Welcome from "../../components/Welcome";
+import { routePaths } from "@/routes/route-paths";
+import { useCurrentUser } from "@/data/user/active-user-query";
+import Welcome from "@/components/Welcome";
 import Skeleton from "react-loading-skeleton";
 import WorkspaceSkeletonCard from "./WorkspaceSkeletonCard";
