+++ src/features/stack-components/detail/StackComponentDetail.tsx
@@ -1,7 +1,7 @@
 import { useParams } from "react-router-dom";
 import { useStackComponentDetail } from "./query";
 import OverviewCard from "./cards/OverviewCard";
-import { ReactComponent as Cube } from "../../../assets/cube.svg";
+import { ReactComponent as Cube } from "@/assets/cube.svg";
 import ConfigCard from "./cards/ConfigCard";
 import StackComponent404 from "./StackComponent404";
